+++ src/main.rs
@@ -14,6 +14,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         Some("send") => doorbell::send_command(),
         Some("control") => doorbell::control_command(),
         Some("pulse") => doorbell::pulse_command(),
+        Some("roster") => doorbell::roster_command(),
         Some("init") => initializer::command(),
         Some("plugin") => plugins::command(),
         Some("toolbox") => toolbox::command(),